add feature logic
