+++ Cargo.toml
@@ -23,7 +23,6 @@ tokio = { version = "1.27.0", features = ["full"] }
 tower-http = { version = "0.5.0", features = ["cors"] }
 tracing = "0.1.40"
 tracing-subscriber = {version = "0.3.18", features = ["json", "time"]}
-rdkafka = { version = "0.36.2", features = ["cmake-build", "ssl"] }
 cron = "0.12.1"
 ndarray = "0.15.6"
 scouter = { version = "= 0.3.0-rc.7" , features = ["dispatch"]}
@@ -31,8 +30,18 @@ metrics-exporter-prometheus = "0.15.1"
 metrics = "0.23.0"
 async-trait = "0.1.81"
 colored = "2.1.0"
-lapin = "2.5.0"
 futures = "0.3.30"
+
+# Kafka dependencies
+rdkafka = { version = "0.36.2", optional = true, features = ["cmake-build", "ssl"] }
+
+# RabbitMQ dependencies
+lapin = { version = "2.5.0", optional = true }
+
+[features]
+default = []
+kafka = ["rdkafka"]
+rabbitmq = ["lapin"]
 
 [dev-dependencies]
 approx = "0.5.1"
