+++ src/consumer/mod.rs
@@ -1,0 +1,3 @@
+pub mod base;
+pub mod kafka;
+pub mod rabbitmq;
