+++ src/consumer/rabbitmq/mod.rs
