+++ src/lib.rs
@@ -1,5 +1,4 @@
 pub mod alerts;
 pub mod api;
-pub mod kafka;
-pub mod rabbitmq;
+pub mod consumer;
 pub mod sql;
