--- src/rabbitmq/startup.rs
@@ -1,35 +1,0 @@
-use crate::rabbitmq::consumer::{start_rabbitmq_background_poll, MessageHandler};
-use crate::sql::postgres::PostgresClient;
-use sqlx::{Pool, Postgres};
-use std::result::Result;
-use tracing::info;
-
-pub async fn startup_rabbitmq(pool: Pool<Postgres>) -> Result<(), anyhow::Error> {
-    info!("Starting RabbitMQ consumer");
-
-    let num_rabbits = std::env::var("RABBITMQ_CONSUMERS_COUNT")
-        .unwrap_or_else(|_| "3".to_string())
-        .parse::<usize>()
-        .map_err(|e| lapin::Error::from(std::io::Error::new(std::io::ErrorKind::Other, e)))?;
-
-    let prefetch_count = std::env::var("RABBITMQ_PREFETCH_COUNT")
-        .unwrap_or_else(|_| "1".to_string())
-        .parse::<u16>()
-        .map_err(|e| lapin::Error::from(std::io::Error::new(std::io::ErrorKind::Other, e)))?;
-
-    for _ in 0..num_rabbits {
-        let rabbit_db_client = PostgresClient::new(pool.clone()).unwrap();
-        let message_handler = MessageHandler::Postgres(rabbit_db_client);
-
-        let rabbit_addr = std::env::var("RABBITMQ_ADDR")
-            .unwrap_or_else(|_| "amqp://guest:guest@127.0.0.1:5672/%2f".into());
-
-        tokio::spawn(async move {
-            start_rabbitmq_background_poll(message_handler, rabbit_addr, prefetch_count)
-                .await
-                .unwrap();
-        });
-    }
-
-    Ok(())
-}
